+++ src/views/transactions/detail/msgs/MsgWithdrawValidatorCommission.vue
@@ -1,0 +1,48 @@
+<template>
+  <BaseTransactionMessage :message="message">
+    <div slot="body">
+      <BaseDetailLink
+        :label="$t('labels.validatorAddress')"
+        :content="validatorAddress"
+        :route="{
+          name: ROUTES.NAME.VALIDATORS_DETAIL,
+          params: { id: validatorAddress },
+        }"
+        look="font-monotype"
+      />
+    </div>
+  </BaseTransactionMessage>
+</template>
+
+<script>
+import BaseDetailLink from '@/components/BaseDetailLink.vue';
+import BaseTransactionMessage from '@/components/BaseTransactionMessage.vue';
+
+import { ROUTES } from '@/constants';
+
+export default {
+  name: 'MsgWithdrawValidatorCommission',
+  description: 'Display a withdraw validator commission transaction message',
+  components: {
+    BaseDetailLink,
+    BaseTransactionMessage,
+  },
+  props: {
+    message: {
+      type: Object,
+      required: true,
+      note: 'Object representing a withdraw validator commission message',
+    },
+  },
+  data() {
+    return {
+      ROUTES,
+    };
+  },
+  computed: {
+    validatorAddress() {
+      return this.message.validator_address || '-';
+    },
+  },
+};
+</script>